+++ airflow/providers/jenkins/sensors/__init__.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,22 +15,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-import unittest
-
-from tests.helm_template_generator import render_chart
-
-
-class WebserverPdbTest(unittest.TestCase):
-    def test_should_pass_validation_with_just_pdb_enabled_v1(self):
-        render_chart(
-            values={"webserver": {"podDisruptionBudget": {"enabled": True}}},
-            show_only=["templates/webserver/webserver-poddisruptionbudget.yaml"],
-        )  # checks that no validation exception is raised
-
-    def test_should_pass_validation_with_just_pdb_enabled_v1beta1(self):
-        render_chart(
-            values={"webserver": {"podDisruptionBudget": {"enabled": True}}},
-            show_only=["templates/webserver/webserver-poddisruptionbudget.yaml"],
-            kubernetes_version='1.16.0',
-        )  # checks that no validation exception is raised