+++ airflow/providers/databricks/CHANGELOG.rst
@@ -19,6 +19,33 @@
 Changelog
 ---------
 
+2.5.0
+.....
+
+Features
+~~~~~~~~
+
+* ``Operator for updating Databricks Repos (#22278)``
+
+Bug Fixes
+~~~~~~~~~
+
+* ``Fix mistakenly added install_requires for all providers (#22382)``
+
+2.4.0
+.....
+
+Features
+~~~~~~~~
+
+* ``Add new options to DatabricksCopyIntoOperator (#22076)``
+* ``Databricks hook - retry on HTTP Status 429 as well (#21852)``
+
+Misc
+~~~~
+
+* ``Skip some tests for Databricks from running on Python 3.10 (#22221)``
+
 2.3.0
 .....