+++ airflow/providers/github/CHANGELOG.rst
@@ -19,6 +19,22 @@
 
 Changelog
 
+1.0.3
+.....
+
+Bug Fixes
+~~~~~~~~~
+
+* ``Fix mistakenly added install_requires for all providers (#22382)``
+
+1.0.2
+.....
+
+Misc
+~~~~~
+
+* ``Add Trove classifiers in PyPI (Framework :: Apache Airflow :: Provider)``
+
 1.0.1
 .....