+++ airflow/providers/samba/CHANGELOG.rst
@@ -19,6 +19,22 @@
 Changelog
 ---------
 
+3.0.4
+.....
+
+Bug Fixes
+~~~~~~~~~
+
+* ``Fix mistakenly added install_requires for all providers (#22382)``
+
+3.0.3
+.....
+
+Misc
+~~~~~
+
+* ``Add Trove classifiers in PyPI (Framework :: Apache Airflow :: Provider)``
+
 3.0.2
 .....