+++ airflow/providers/sftp/CHANGELOG.rst
@@ -19,6 +19,22 @@
 Changelog
 ---------
 
+2.5.2
+.....
+
+Bug Fixes
+~~~~~~~~~
+
+* ``Fix mistakenly added install_requires for all providers (#22382)``
+
+2.5.1
+.....
+
+Misc
+~~~~~
+
+* ``Add Trove classifiers in PyPI (Framework :: Apache Airflow :: Provider)``
+
 2.5.0
 .....