+++ airflow/providers/snowflake/CHANGELOG.rst
@@ -19,6 +19,27 @@
 Changelog
 ---------
 
+2.6.0
+.....
+
+Features
+~~~~~~~~
+
+* ``Add support for private key in connection for Snowflake (#22266)``
+
+Bug Fixes
+~~~~~~~~~
+
+* ``Fix mistakenly added install_requires for all providers (#22382)``
+
+2.5.2
+.....
+
+Misc
+~~~~
+
+* ``Remove Snowflake limits (#22181)``
+
 2.5.1
 .....